Remove NavLink
