--- packages/commons-ui-core/src/NavList/NavMenu.js
@@ -1,83 +1,0 @@
-import { Grid, Typography } from "@mui/material";
-import Button from "@mui/material/Button";
-import { useTheme, styled } from "@mui/material/styles";
-import useMediaQuery from "@mui/material/useMediaQuery";
-import PropTypes from "prop-types";
-import * as React from "react";
-
-const NavTypography = styled(Typography)({
-  "&:hover": {
-    textDecoration: "underline",
-    color: "blue",
-  },
-});
-
-const NavContainer = styled(Grid)({
-  background: "white",
-});
-
-const NavButton = styled(Button)({
-  boxShadow: "none",
-  textTransform: "capitalize",
-  fontSize: 16,
-  color: "black",
-  margin: 16,
-  backgroundColor: "transparent",
-  borderColor: "none",
-  "&:hover, &:active, &:focus, &:focus-within": {
-    backgroundColor: "transparent",
-    borderColor: "none",
-    boxShadow: "none",
-  },
-});
-
-function NavMenu({ children, menu, buttonVariant, typographyVariant, footer }) {
-  const theme = useTheme();
-  const isMobile = useMediaQuery(theme.breakpoints.down("md"));
-
-  if (!menu?.length) {
-    return null;
-  }
-
-  return (
-    <NavContainer
-      container
-      direction={isMobile || footer ? "column" : "row"}
-      alignItems="flex-start"
-      justifyContent={isMobile ? "flex-start" : "flex-end"}
-    >
-      {menu.map((item) => (
-        <Grid item key={item.label}>
-          <NavButton variant={buttonVariant} href={item.href}>
-            <NavTypography variant={typographyVariant}>
-              {item.label}
-            </NavTypography>
-          </NavButton>
-        </Grid>
-      ))}
-      {children}
-    </NavContainer>
-  );
-}
-
-NavMenu.propTypes = {
-  buttonVariant: PropTypes.string,
-  typographyVariant: PropTypes.string,
-  menu: PropTypes.arrayOf(
-    PropTypes.shape({
-      label: PropTypes.string,
-      href: PropTypes.string,
-    }).isRequired
-  ).isRequired,
-  children: PropTypes.node,
-  footer: PropTypes.bool,
-};
-
-NavMenu.defaultProps = {
-  footer: undefined,
-  buttonVariant: undefined,
-  typographyVariant: undefined,
-  children: undefined,
-};
-
-export default NavMenu;
